+++ find_film.txt
@@ -1,1 +1,1 @@
-1999 комедия 5.8 10 f
+2010 комедия 5.8 10
