+++ list_films.txt
@@ -1,6 +1,7 @@
-Л 1999 комедия 5.5
-р 2000 комедия 55.5
-g 1999 комедия 6
-е 1999 ужасы 6.1
-н 1999 комедия 7
-в 1999 комедия 11
+Елки 2010 комедия 6.9
+Призрак 2005 комедия 5.5
+Последний богатырь 2017 фэнтези 8
+Крик 2022 ужасы 6.1
+Наследники 2015 мюзикл 7.9
+Тайна Коко 2017 детский 7.1
+Жизнь 2012 драмма 15
